seperated the widget from the propagator
