+++ src/uncertainty_propagator/__init__.py
@@ -1,1 +1,2 @@
-from .propagator import Propagator, VariableInputRow
+from .propagator import Propagator
+from .jupyter_widget import VariableInputRow, PropagatorWidget
